requirements.txt
# データサイエンス系ライブラリ
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.0.0

# NLP関連
nltk>=3.8.0
fugashi>=1.2.0
ipadic>=1.0.0

# Webフレームワークとユーティリティ
streamlit>=1.25.0
requests>=2.30.0
python-dotenv>=1.0.0

# データベースとORM
sqlalchemy>=2.0.0

# その他依存ライブラリ
tqdm>=4.65.0
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.0.0
requests==2.32.4
six==1.17.0
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
